+++ apis/validators/validate_test.go
@@ -23,137 +23,12 @@ import (
 	"github.com/knative/pkg/apis"
 )
 
-type foo struct {
-	Default string `validate:"-"`
-}
-
-type foo_k8s struct {
-	Default      string `json:"default,omitempty"`
-	OptionalName string `json:"optionalName" validate:"QualifiedName"`
-	RequiredName string `json:"requiredName" validate:"QualifiedName;Required"`
-}
-
-type non_json_k8s struct {
-	OptionalName string `validate:"QualifiedName"`
-	RequiredName string `validate:"QualifiedName;Required"`
-}
-
-type complex_types struct {
-	Name         string
-	RequiredName string `validate:"Required"`
-	Num          int
-	RequiredNum  int `validate:"Required"`
-	Foo          foo
-	RequiredFoo  foo `validate:"Required"`
-}
-
 const invalidQualifiedNameError = `name part must consist of alphanumeric characters, '-', '_' or '.', and must start and end with an alphanumeric character (e.g. 'MyName',  or 'my.name',  or '123-abc', regex used for validation is '([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9]')`
 
 type ValidateTest struct {
 	name string
 	obj  interface{}
 	want *apis.FieldError
-}
-
-func TestValidate(t *testing.T) {
-	tests := []ValidateTest{{
-		name: "default",
-		obj: foo{
-			Default: "default",
-		},
-		want: nil,
-	}, {
-		name: "valid k8s",
-		obj: foo_k8s{
-			Default:      "default",
-			OptionalName: "valid",
-			RequiredName: "valid",
-		},
-		want: nil,
-	}, {
-		name: "missing required k8s name",
-		obj:  foo_k8s{},
-		want: &apis.FieldError{
-			Message: `missing field(s)`,
-			Paths:   []string{"requiredName"},
-		},
-	}, {
-		name: "invalid optional k8s name",
-		obj: foo_k8s{
-			Default:      "default",
-			OptionalName: "v@lid",
-			RequiredName: "valid",
-		},
-		want: &apis.FieldError{
-			Message: `invalid key name "v@lid"`,
-			Paths:   []string{"optionalName"},
-			Details: invalidQualifiedNameError,
-		},
-	}, {
-		name: "invalid required k8s name",
-		obj: foo_k8s{
-			RequiredName: "v@lid",
-		},
-		want: &apis.FieldError{
-			Message: `invalid key name "v@lid"`,
-			Paths:   []string{"requiredName"},
-			Details: invalidQualifiedNameError,
-		},
-	}, {
-		name: "invalid optional and required k8s names",
-		obj: foo_k8s{
-			OptionalName: "val!d",
-			RequiredName: "v@lid",
-		},
-		want: (&apis.FieldError{
-			Message: `invalid key name "val!d"`,
-			Paths:   []string{"optionalName"},
-			Details: invalidQualifiedNameError,
-		}).Also(&apis.FieldError{
-			Message: `invalid key name "v@lid"`,
-			Paths:   []string{"requiredName"},
-			Details: invalidQualifiedNameError,
-		}),
-	}, {
-		name: "non-json invalid optional and required k8s names",
-		obj: non_json_k8s{
-			OptionalName: "val!d",
-			RequiredName: "v@lid",
-		},
-		want: (&apis.FieldError{
-			Message: `invalid key name "val!d"`,
-			Paths:   []string{"OptionalName"},
-			Details: invalidQualifiedNameError,
-		}).Also(&apis.FieldError{
-			Message: `invalid key name "v@lid"`,
-			Paths:   []string{"RequiredName"},
-			Details: invalidQualifiedNameError,
-		}),
-	}, {
-		name: "complex types",
-		obj: complex_types{
-			RequiredName: "foo",
-			RequiredNum:  42,
-			RequiredFoo: foo{
-				Default: "hi",
-			},
-		},
-		want: nil,
-	}, {
-		name: "empty complex types",
-		obj:  complex_types{},
-		want: (&apis.FieldError{
-			Message: `missing field(s)`,
-			Paths:   []string{"RequiredName"},
-		}).Also(&apis.FieldError{
-			Message: `missing field(s)`,
-			Paths:   []string{"RequiredNum"},
-		}).Also(&apis.FieldError{
-			Message: `missing field(s)`,
-			Paths:   []string{"RequiredFoo"},
-		}),
-	}}
-	doTestValidate(t, tests)
 }
 
 func doTestValidate(t *testing.T, tests []ValidateTest) {
